backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
pydantic
python-dotenv
